Updated ColorExtensions.Difference to use the 3D difference between colors in the RGB color volume instead of just summing the difference of the RGB color components.

+++ CyberCavsScoutingSystem/GameMakerWpf/Validation/Data/AllianceValidationData.cs
@@ -97,7 +97,7 @@ public static class AllianceValidationData {
 		private const int SimilarityWarningThreshold = 10;
 		private const int SimilarityAdvisoryThreshold = 20;
 
-		public static Optional<Error> GetColorSimilarityError(int colorDifference, string otherAllianceName) {
+		public static Optional<Error> GetColorSimilarityError(double colorDifference, string otherAllianceName) {
 
 			return colorDifference switch {
 				<= SimilarityErrorThreshold => new Error("Identical Color", ErrorSeverity.Error,
